+++ deploy/sample-backup-configs/sample_config_build.ini
@@ -23,6 +23,8 @@ firesim-boom-singlecore-nic-l2-llc4mb-ddr3
 # SHA3 configs for tutorial
 # firesim-singlecore-sha3-no-nic-l2-llc4mb-ddr3
 # firesim-singlecore-sha3-print-no-nic-l2-llc4mb-ddr3
+test-vitis
+
 
 [agfistoshare]
 firesim-rocket-quadcore-nic-l2-llc4mb-ddr3
