+++ deploy/sample-backup-configs/sample_config_build_recipes.ini
@@ -19,6 +19,18 @@
 # # BUILD PLATFORM SPECIFIC (if unspecified, use the defaultbuildhost)
 # buildhost=defaultbuildhost
 
+[test-vitis]
+# # DESIGN SPECIFIC
+DESIGN=FireSim
+TARGET_CONFIG=FireSimRocketConfig
+deploytriplet=None
+# # RUN PLATFORM SPECIFIC
+PLATFORM_CONFIG=BaseVitisConfig
+postbuildhook=
+fpgaplatform=VitisBitBuilder
+# # BUILD PLATFORM SPECIFIC
+buildhost=localbuildhost
+
 # Quad-core, Rocket-based recipes
 [firesim-rocket-quadcore-nic-l2-llc4mb-ddr3]
 DESIGN=FireSim
