+++ deploy/sample-backup-configs/sample_config_runfarm.ini
@@ -1,0 +1,22 @@
+[ec2-f1-runfarm]
+# debate on not having the class name exposed
+#provider=ec2
+providerclass=EC2RunFarm
+runfarmtag=mainrunfarm
+f1_16xlarges=0
+m4_16xlarges=0
+f1_4xlarges=0
+f1_2xlarges=1
+runinstancemarket=ondemand
+spotinterruptionbehavior=terminate
+spotmaxprice=ondemand
+
+[local-runfarm]
+#provider=ipaddr
+providerclass=IpAddrRunFarm
+# flag to control where the simulation runs out of
+#remotebuilddir=
+ipaddr=localhost
+#need something to tell what the instance is (is it a vitis instance or is it a F1 instance)
+# this is the node class to use (can use an F1 instance instead)
+runinstancedescriptor=Vitis_Inst
