+++ deploy/sample-backup-configs/sample_config_runtime.ini
@@ -3,7 +3,7 @@
 # documentation of all of these params.
 
 [runfarm]
-runfarmtag=mainrunfarm
+runfarm=local-runfarm
 
 f1_16xlarges=1
 m4_16xlarges=0
